Use blubberoid.wikimedia.org to process blubber.yaml
Simplified `Blubber` client to handle only the transcompilation of Blubber
configuration into Dockerfile and not the actual building of images.
Building was moved into `PipelineRunner` where all other Docker commands
are invoked.

Blubber client was refactored to use the WMF production deployment of
Blubberoid instead of relying on a locally installed binary of the
`blubber` CLI.

Bug: T212247
Change-Id: Ib403786af7af6ce9d469798452da512fa535f2b4