pipeline: Support Zuul based SCM for configure stage
The configure stage previously depended on the `scm` job property being
available. This is only available, however, for jobs that source a
`Jenkinsfile` from the repo, not the pipeline scripts sourced from
`integration/config` groovy files.

Calling `PipelineBuilder#build` in the context of the latter results in
the following error:

> ERROR: 'checkout scm' is only available when using "Multibranch
> Pipeline" or "Pipeline script from SCM"

The builder needs to support being called from both a `Jenkinsfile` and
the scripts defined in integration/config.

Change-Id: I1dae576c35c4ea1a8746ec418290c52679314598