pipeline: Work around Groovy CPS replaceAll limitations
Groovy CPS does not correctly implement `replaceAll(Pattern, Closure)`.
Specifically, in cases where the pattern contains a capture group, it
does not pass each match for each capture group to the given closure as
a separate argument, as Groovy's implementation does.

The workaround in this case is to use a substring operation to omit the
surrounding `${}` from the matched variable expression when
interpolating string values against a node context.

Change-Id: I23dcceeb416d5c22ad0f36ab3dca8303a9ed7fd4