Default to localhost for JENKINS_HOST in Makefile
The original `Makefile` was written assuming the Docker daemon is
accessible over TCP and so the `JENKINS_HOST` can be inferred from
`DOCKER_HOST`. That isn't usually the case in a dev environment, so it
makes more sense to assume the host will be `localhost`.

Some fixes to the system-test environment are also included, a fix to
the group permissions of the docker-in-docker bind-mounted unix socket,
and updates to the blubber configuration for each test service.

Change-Id: Ib57440a0095c7987b3d014517a1495e4ba3ee021