Require a `pipelinelib` node label when scheduling
Including a base `pipelinelib` node label will allow for a greater
degree of control over the pool of executors available for executing
pipelinelib based jobs. Additionally, it will ensure that node labels
are never empty which prevents Jenkins from scheduling pipelinelib based
jobs in some cases (an empty set of stages for a pipeline for example).

Change-Id: Id2e44e8990623c6177ae2c4bfc173c15b7ab7172