pipeline: Prevent running of containers with arbitrary image refs
The `PipelineRunner.run` method, and by extension `PipelineStage.run`,
is only meant to run images by ID (SHA256 digest) which are images known
locally to the Docker daemon. Since the image ID argument is not
validated and is passed directly to `docker run` which allows for any
valid image ref, it was previously possible to configure a pipeline
stage thusly to run containers from any valid Docker registry.

  pipelines:
    foo:
      stages:
        - name: mine
          run:
            image: registry.mine/crypto-miner:unstable
            arguments: [money, money, money]

To prevent such abuse, qualify the image ID argument passed to
`PipelineRunner.run` with a "sha256:" prefix before passing to `docker
run` command. This will ensure the reference is interpreted only as an
image ID (digest) and any pull made to the registry—should the image not
exist—will fail.

Change-Id: Ic4bba4ff2bad6ed0a35996f124640801bf832bc3