Enable deployment-charts update for published image
Adds a promote step to update the helm chart for a service when a new
image is published. Multiple charts can be updated for one service.
Example:

promote:
  - chart: "foo"
    environments:
      - "staging"

promote: true
will default to the following, which will attempt to update a chart with
the name of the project for all environments:

promote:
  - chart: "project-name"
    environments: []
    version: "${.imageTag}"

Bug: T255835
Depends-on: I14a83090ec35e8116786ab61e527951a2b46f406
Depends-on: I69fe38554e58d27ea0d46d35da9c4490e124c246
Change-Id: Id8232dd3fdc1a986ae24ab821e2778b75779c618