Add Helm chart repo to pipeline
WMF now hosts a chart registry outside of the deployment-charts git
repo. All pipeline config files will need to update their chart
locations.

Instead of requiring chart url in the config file, this adds the chart
registry url as a variable in PipelineStage. Config files should now
only specify the chart name and version in the deploy block if desired.
If not specified, the following defaults will be applied:

chart:
  name: 'project-name'
  version: ''

helm does not require a chart version to install. If no version is
specified, the latest version will be used.

Bug: T261346
Change-Id: Ifeb5caec6bf120b86ae16afd0315d98ffbc63dad