Refactor promote step to use withCredentials
git credential cache must be initiated in the target directory and
then the password must be typed in. This requires a refactor of
the promote step. Use withCredentials to store the password in git
credential helper and then un-set it for each push.

After checkout into the target directory, it is necessary to cd into
that directory because the checkout does not do it automatically.

Add git config user.email and user.name

Change-Id: I4e400bbbc4d31609c336d7739e9b9d200781d001