Allow for standard docker pushes with authentication
Introduces a `registryPushMethod` property to `PipelineRunner` that
allows for two possible values: "wmf-pusher" and "docker-push". The
former (default) configures the runner to use the `docker-pusher`
wrapper script for pushing images to the official WMF registry. The
latter configures the runner to use a standard `docker push` command,
allowing for pushes to other registries as well.

If "docker-push" is used, another new property `registryCredential` must
be set with the ID of a Jenkins credential that has push access to the
remote registry. This method requires a new
`docker-credential-environment` credential helper to be installed on the
Jenkins node.

Depends-On: I71597ece773f5e6eeb04397a74ee4637b89d14b3
Bug: T265177
Change-Id: Ib2ae720f8f64ced2df8f86315c6e6db25feaebff