Expose Jenkins job parameters and Git checkout info
Bind all available job parameters to the setup context for use in user
defined pipelines as well as any information returned by
`checkout(scm)`. Normalize some of the latter as:

 - commit: The Git commit of the checked out patchset
 - branch: The Git branch used to check out the patchset
 - remote: The Git remote URL used to check out the patchset

In addition, include more information in the job output about what data
is exported by each stage. This should make it easier for users to know
what can be referenced in subsequent stages and help with debugging.

Change-Id: I197e63bcd483906b6a953770e814d5a4643cca67