PipelineRunner: Remove imageName from release name
Using the image name in the release name can cause the release name to
be very long. It's used in the helm charts as part of the deployment
name and requires truncation. Ideally this should happen in the
helm chart, but since the chart name is usually the same as the image
name, and the image name can be discovered by describing the pod or
deployment, let's also remove it from the release name when doing ci
deployments.

Change-Id: I4beddfc978b2dd26b0972863f5c83cae789459a4