Namespace build parameters to `setup.params.[name]`
Currently, if a job defines a parameter that conflicts with any of the
internally defined exports, it will cause an
`ExecutionContext.NameAlreadyBoundException`. For example, a job that I
created on releases-jenkins used a parameter called "branch" which
conflicts with the "branch" export set to the git branch returned by
`checkout`.

Namespacing the build parameters exported by setup avoids such
collisions.

Note that `NodeContext#getAt` had to be refactored to work with periods
in the second part of variable expressions.

Change-Id: Ia3f1b8233ed60a383a1907f2ee2347f24ffb20e5