Export container process output from run step
Introduces a `run.tail` field to stage configuration that allows users
to specify a number of trailing lines to save from the container's
output. The default value is 0 as saving output by default for all run
containers would bloat memory usage, CPS serialization, and console
output.

Adds an `.output` export to the run step which will contain the given
trailing lines from the stdout of the container process. This can be
useful in chaining inputs/outputs between steps and stages.

Refactors `PipelineRunner` to support capturing the output of executed
containers. Removed the use of `exec` in favor of catching
`InterruptedException`. Both ensure proper cleanup (stopping of
container process) should Jenkins pass a TERM signal to the shell
process. However, the latter allows us to redirect stdout of the `docker
run` process using `tee` to simultaneously capture it as a return value
and continue outputting it to the Jenkins console.

Additionally, `PipelineRunner#build` was refactored to use `--iidfile`
instead of parsing output for the resulting image ID as the latter is
unreliable. As a result, users may now see the build output in the
console.

Change-Id: I1d790045c73c1540ab30556c662bed8958eb3429