PipelineRunner: allowedCredentials
Adds an allowedCredentials property to the PipelineRunner for defining
allowed credentials per pipeline.

Adds a PipelineCredentialsManager to manage credential bindings and env
var string generation. Checks against the allowedCredentials map to
determine whether credentials can be passed to withCredentials in the
'run' method.

The allowedCredentials should be in the form of [credentialId:
"bindingType"]

Bug: T269902
Change-Id: I1114d82293f263fcf3d57d80b9d803e052255c11