Add helm deploy timeout override
Some charts might take longer for all pods to be ready than is afforded
by the default timeout setting. This adds an optional 'timeout' property
to the deploy stanza, which users can define in their
.pipeline/config.yaml.

Additional changes 1) Fix a bug that would default the image name for helm
deployment to a combination of image name and tag instead of image name
only, and 2) Set a default image tag if none is provided.

Bug: T277125
Change-Id: I94359be83f8ac9b9a43bbbf8e6dcc360623565cc