Support `copy` for copying files from containers to the local context
Refactored `run` step to save the generated container name in a context
variable `[stage].container` and delay removal of the stopped container
until teardown, allowing subsequent access to the container filesystem.

Implemented `copy` configuration, allowing shorthand and longhand forms.

Example:

    copy:
      - from: '${foo.container}'
        source: 'source/foo'
        destination: 'dest/foo'

The above copies the file `source/foo` from the container that ran in
the `foo` stage to `dest/foo` in the local context directory.

Example (shorthand):

    copy: [source/foo, source/bar]

The above specifies only source files and is expanded to the following.

    copy:
      - from: '${.container}'
        source: 'source/foo'
        destination: 'source/foo'
      - from: '${.container}'
        source: 'source/bar'
        destination: 'source/bar'

The `from` field defaults to the container that ran in the current stage
and `destination` defaults to the source path.

Additionally, stubbed `publish.files` implementation along with
documentation was removed to avoid confusion.

Bug: T277777
Change-Id: Id99fedb9037fc08d32370b5122bbafbc38372053