Always copy the contents of source directories
As the inline comment reads, the logic that `docker cp` employs when it
comes to source directories (whether to copy its contents or the
directory itself into the destination) is truly heinous. We're
normalizing the behavior to always copy source directory _contents_ by
appending a `.`. The source is assumed to be a directory if it is
terminated with a `/`.

For the record, the actual logic of `docker cp` for source directories
which we're abstracting away is as follows.[1]

 - DEST_PATH does not exist
   - DEST_PATH is created as a directory and the contents of the source
     directory are copied into this directory
 - DEST_PATH exists and is a file
   - Error condition: cannot copy a directory to a file
 - DEST_PATH exists and is a directory
   - SRC_PATH does not end with /. (that is: slash followed by dot) the
     source directory is copied into this directory
   - SRC_PATH does end with /. (that is: slash followed by dot) the
     content of the source directory is copied into this directory

[1]: https://docs.docker.com/engine/reference/commandline/cp/

Change-Id: I3a38a4c0103e2876dc1c6b4812b7e48820e57022