Support email notification upon stage failure
Support configuration at the pipeline and stage scope that will result
in email notification upon stage failure.

Example config:

    pipelines:
      foo:
        notify:
          email:
            to: engprod@lists.wikimedia.org
        stages:
          - name: normal-stuff
          - name: security-patches
            build: apply-patches
            run: true
            notify:
              email:
                to: security-team@wikimedia.org
                subject: Security patch application failed

In the example configuration, the pipeline level configuration indicates
that failure at any stage should result in a notification to the EngProd
team with the default subject and email body. The stage level
configuration for `security-patches` overrides some of this to customize
the recipient (Security Team) and subject line.

Additionally a CPS related bug was fixed in the raising of
`NameNotFoundException` exceptions in `ExecutionContext$NodeContext`.
The `getAllKeys` method must be called outside of the exception
constructor as it is CPS transformed. The implementation of `getAllKeys`
was also fixed and a test was added.

Change-Id: If3b16e30c0a8858dcb4b9eff3c9aa103d489405d