Add 'trigger' step
The 'trigger' step triggers a downstream job on the local Jenkins
server. This is the final step that occurs in a stage.

Security notes:

* With this change, any user with +2 permission on a repo can trigger
any jenkins job listed as allowed for the pipeline as passed to the
PipelineBuilder. This is currently configured in the integration/config
repo for most repositories.

Change-Id: I74df896bff62a1268ab33dfea7ad8f7db420ef94