Enforce pipefail on all run step commands
In instances where `withOutput` is used, a non-zero exit status from the
run-step command did not result in an exception as is normally the case.
This is because `withOutput` appends the command string with `| tee ...`
to capture output to a file. Enforcing `pipefail` on all shells used to
execute run-step commands fixes the issue.

Refactored the static command prefix to a closure set as a
`PipelineRunner` property. This makes it customizable by the user. Also
added an explicit `bash` interpreter as this is required when using
`pipefail`.

Bug: T290608
Change-Id: I5f6171a58e9dc353a049d891e215c2d7b901e23f