Replace regex based variable parser with a lexer
The previous variable parser relied on `replaceAll` and a regular
expression to match and replace variable expressions within user
provided strings. This implementation does not appear to be CPS
compatible and fails if Jenkins attempts to serialization a thread's
state during variable replacement. When a serial execution graph is used
(the default), this is not likely to happen. However, if the user
provides an execution graph with a higher number of concurrent branches,
the case is reached often and results in a
`java.io.NotSerializableException` exception within the CPS serializer.

Replacing the regex based parser with a string based lexer avoids this
case.

The `Utility.randomAlphanum` static method was also refactored and
marked `@NonCPS` to avoid a similar CPS serialization issues related to
its `IntRange` and character range.

See JENKINS-56343 and JENKINS-63074 for related Jenkins issues.

Added a systemtest that can intermittently reproduce the CPS
serialization bug using four independent execution branches and one
stage that completes much earlier than the other three.

Bug: T287507
Change-Id: I9d4b965e2505ad5cb08d148cf0ec4038a7dd0586