Perform validation using JSON schema and ajv-cli
Implemented validation in `PipelineBuilder` that runs `ajv` against the
user provided configuration using the JSON schema before attempting to
fully parse the configuration into `Pipeline` objects.

Note that for some really strange and unknown reason, the stdout of the
`ajv` container does not appear in the console during systemtest runs.
This may have something to do with the docker-in-docker environment (the
socket is bind mounted to allow the Jenkins instance to run docker via
the host's dockerd) but it's not 100% clear. The problem does not occur
when run directly on the host.

Added missing `imagePullPolicy` build step schema.

Bug: T225335
Change-Id: I53886006a99e9c71ef8246e65c435cfa9a3c2e2a