Ignore non-pipeline variable tokens in config
With the refactoring of variable substitution, tokens such as `${FOO}`
that represent non-pipeline variables (i.e. shell environment variables)
were being resolved against the current `NodeContext` binding and would
result in `NameNotFoundException` exceptions. This is a regression as
the previous implementation would ignore such patterns (variables not
containing a `.`).

Fix the regression by checking for a lack of `.` in the variable name
when catching `NameNotFoundException` and swallowing the exception.

Bug: T291554
Change-Id: I6b2c746f099a7314d5dad475324d75c5729260f7