Archive directories as tar.gz when `copy.archive` is true
When `copy.destination` is a directory and `copy.archive` is true, first
create a tar.gz from the directory before archiving it as a Jenkins
artifact.

Expose all artifacts created by `copy.archive: true` within a stage as a
`artifactURLs` variable that can be easily used by downstream actions or
stages. The entries within the variable are newline separated, making it
easy to process via `$each` list comprehensions or use as a single URL
when there is only one copy operation.

Example use case: publish documentation via a downstream job

    stages:
      - name: publish-docs
        build: documentation
        run: true
        copy:
          - source: /srv/app/build/docs/groovydoc
            destination: docs
            archive: true
        trigger:
          name: publish-artifact-to-doc-wikimedia-org
          parameters:
            ARTIFACT_URL: '${.artifactURLs}'
            WMF_CI_PUB_DEST: 'my/project/doc/root'

Bug: T308591
Change-Id: I6467f5843a0c9b89a835393a461ec158922251ba