Fix blubber file header processing
The previous fix to avoid groovy-cps's incorrect implementation of
`takeWhile` replaced the latter with a loop that processed all commented
lines in the file, not just the first set. As a result, any config with
comments after the header section have leading lines subsequently
removed before the build.

Use a for loop with a break condition to correctly refactor the original
`takeWhile`.

Follows up Ib2d42449697b4383a2272db622f70964c3fed23d

Bug: T320953
Bug: T318290
Change-Id: I6cd881b1dd10ee32e9e8177894e111d266611266