Allow more BuildKit frontend image names (v2)
As far as I can tell, the image was renamed between v0.12.0 and v0.12.2.
Keep the old name and version as the default, to avoid breaking anything
at the moment, but allow the new name in the syntax directive as well.

This is a resubmit of change Ie0397544af (commit 702ecfbc64), which had
to be reverted due to T331497; to avoid that issue happening again,
annotate parseImageRef() (which we now call from a constructor) as
@NonCPS. (Note that randomAlphanum(), directly below, was already
annotated this way, probably for a similar reason.)

See https://wiki.jenkins.io/JENKINS/Pipeline-CPS-method-mismatches.html#PipelineCPSmethodmismatches-Constructors

Bug: T329553
Change-Id: I72f8f42ab8a6b42dda7cbf8eef08a1ffa5434e27